test(api): move shared scheduler helpers into common-utils
